add subcommand: template
Note: this is a partially generic build, without the linting and testing.
It templates any given file ext in any given directory, not just .yaml files
in backend/x.

No doubt build could be reimplemented on top of it, but for now there's a bit
of duplication between the two subcommands.