De-bundle linting/testing from build and deploy commands
  build: remove linting, testing and exec

  deploy: remove linting

  exec: add

Note that with the build and execution steps (testing, deployment, etc.) being
separated, there's more risk that you may execute accidentally against the
wrong environment or service. To mitigate this, a JSON file is created during
the build which contains environment variables for the environment and service.
The `exec` and `deploy` commands check this file first to confirm the built
environment/service is the same as those specified in the command.